Animals now die of starvation, working on balance issues
Animals had a tendency to move to the bottom of the grid - this issue
has been mostly mitigated, though now they have a much smaller tendency
to move to either the top or the bottom. Need to come up with a way to
place weights on each type of wildlife when spawning so that a
significantly larger amount of trees and grass can be generated.
Currently, all of the insects are dying off first because there are not
enough flora to feed them.